backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
pydantic[email]
python-multipart
pypdf2
httpx
groq